Move score loading functions to file_utils
